refactor: decompose CapturePage into focused single-bin and shared modules
Extract CameraStartScreen, Viewfinder, ShutterButton, ThumbnailItem, and
StatusBadge into standalone files; split single-bin flow into
CaptureSingleBinPage; slim useCapture by removing unused setPhotoGroupId,
facingMode, and stopCamera exports.
